backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
google-generativeai==0.8.0
supabase==2.7.4
python-multipart==0.0.9
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
